Changement de type pour autoriser comparaison

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: gd-harco <gd-harco@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/11 12:30:19 by gd-harco          #+#    #+#             */
-/*   Updated: 2022/11/11 14:15:09 by gd-harco         ###   ########lyon.fr   */
+/*   Updated: 2022/11/12 12:39:09 by gd-harco         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	x;
+	size_t	x;
 
 	x = 0;
 	while ((s1[x] || s2[x]) && x <= n)
